refactor(marketing-analytics): extract compare pivot to shared base method
